Theme System: Split & organised tests, changed module version to string
